Attack of the clones (#8)
* Shoot a blade from the robot to the player when it's visible

* Fix robot suicide by storing shooter id for each blade

* Add floating point number comp utility

* Enhance robot shooting system

- add robot shooting animation
- add character state map
- check if robot faces toward player when shooting

* Check blade-player collision

* Add robot to level config

+++ COMP4300/EntityManager.h
@@ -67,7 +67,7 @@ namespace goldenhand
     private:
         EntityVector mEntities;
         EntityMap mEntityMap;
-        size_t mTotalEntities;
+        uint16_t mTotalEntities;
     
         EntityVector mToAdd;
